requirements.txt
openai>=1.30.0
streamlit>=1.35.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
sentence-transformers>=2.2.0